Type checking implementation
New class Metainfo:
* Contains only one field at the moment: The type(int/bool) that the next
* expression has to have to be syntactical correct
e.g. when the expression is "int x; x:= 3" then a Metainfo-object is created after
the identifier "x" was read. This object will contain the expected type
TYPE_INT because "x" was declared as int. The Metainfo-object is later used
to determine if the right side of the expression "3" is of the right type.